Change attackCooldown back to f16
A full f32 is unneeded for this operation.

+++ inc/player.h
@@ -39,7 +39,7 @@ struct _Player {
   V2f32 position;
   Sprite* sprite;
   f16 bankDirection;
-  f32 attackCooldown;
+  f16 attackCooldown;
   f32 damageCooldown;
   u8 health;
 };
